tp: split access-path selection out of dataframe bytecode emission
QueryPlanBuilder decided the access path and emitted bytecode in a single
pass, so the two were inseparable. Choosing a filter strategy meant reaching
into register allocation, scratch lifetimes and null plumbing; and the row
estimates those decisions depend on were mutated as a side effect of
emitting each instruction. Nothing could consume the planner's choices
without also consuming the interpreter.

Move everything mechanical into BytecodeLowering: register allocation and
caching, scratch management, materializing a range into an index list,
pruning nulls, translating sparse-null indices, row layout buffers, and the
running row-count and cost estimates. Each Emit* method takes one strategy
the planner has already chosen and expands it into the bytecodes it needs.

QueryPlanBuilder is left with the decisions: filter ordering, which strategy
serves each constraint, whether an index is worth using, and whether a sort
can be skipped.

Splitting the code is not enough on its own, because the vocabulary those
decisions are written in lived in the interpreter. NonNullOp, RangeOp,
NonStringType and the rest are plain subsets of Op and StorageType with
nothing to do with bytecode, so move them next to the types they refine in
core/common. Interpreter code refers to them unqualified and is unaffected.
The planner now names no interpreter type to reach a decision; what it still
touches is the plan artifact itself.

This is a pure restructure, and a step towards describing a query as a
sequence of logical operations that something other than the bytecode
interpreter could consume. Verified byte-for-byte identical: across three
traces, all 1753 serialized query plans -- bytecode, register numbering,
cost and row estimates alike -- are unchanged.
