+++ src/trace_processor/core/dataframe/query_plan.h
@@ -51,6 +51,7 @@
 namespace perfetto::trace_processor::core::dataframe {
 
 class Dataframe;
+class BytecodeLowering;
 
 // Specification for initializing a register before bytecode execution.
 // The plan contains abstract references (column indices, index IDs), and
@@ -228,16 +229,14 @@ struct QueryPlanImpl {
   base::SmallVector<RegisterInit, 16> register_inits;
 };
 
-// Builder class for creating query plans.
+// Chooses the access path for a query.
 //
-// QueryPlans contain the bytecode instructions and interpreter configuration
-// needed to execute a query.
+// This class makes all the decisions a query plan involves: the order filters
+// are applied in, which strategy serves each one, whether an index is worth
+// using, and whether a sort can be skipped. It makes no bytecode of its own;
+// each decision is handed to BytecodeLowering, which knows how to emit it.
 class QueryPlanBuilder {
  public:
-  // Represents register types for holding indices.
-  using IndicesReg = std::variant<interpreter::RwHandle<Range>,
-                                  interpreter::RwHandle<Span<uint32_t>>>;
-
   static base::StatusOr<QueryPlanImpl> Build(
       uint32_t row_count,
       const std::vector<std::shared_ptr<Column>>& columns,
@@ -249,57 +248,7 @@ class QueryPlanBuilder {
       uint64_t cols_used);
 
  private:
-  // Indicates that the bytecode does not change the estimated or maximum number
-  // of rows.
-  struct UnchangedRowCount {};
-
-  // Indicates that the bytecode is a non-equality filter.
-  struct NonEqualityFilterRowCount {};
-
-  // Indicates that the bytecode is a equality filter with given duplicate
-  // state and estimated distinct-value count (0 = unknown).
-  struct EqualityFilterRowCount {
-    DuplicateState duplicate_state;
-    uint32_t estimated_distinct = 0;
-  };
-
-  // Indicates that the bytecode is an IN filter over a value list. Unlike a
-  // scalar equality, an IN matches multiple distinct values; since the list
-  // size is not known at plan time (the RHS may be a subquery), we assume it
-  // selects a fixed number of distinct values. `estimated_distinct` is the
-  // per-column distinct-value count (0 = unknown).
-  struct InFilterRowCount {
-    DuplicateState duplicate_state;
-    uint32_t estimated_distinct = 0;
-  };
-
-  // Indicates that the bytecode produces *exactly* one row and the estimated
-  // and maximum should be set to 1.
-  struct OneRowCount {};
-
-  // Indicates that the bytecode produces *exactly* zero rows and the estimated
-  // and maximum should be set to 0.
-  struct ZeroRowCount {};
-
-  // Indicates that the bytecode produces `limit` rows starting at `offset`.
-  struct LimitOffsetRowCount {
-    uint32_t limit;
-    uint32_t offset;
-  };
-  using RowCountModifier = std::variant<UnchangedRowCount,
-                                        NonEqualityFilterRowCount,
-                                        EqualityFilterRowCount,
-                                        InFilterRowCount,
-                                        OneRowCount,
-                                        ZeroRowCount,
-                                        LimitOffsetRowCount>;
-
-  // Constructs a builder for the given indices and columns.
-  // cache is used for caching column/index registers.
-  QueryPlanBuilder(interpreter::BytecodeBuilder& builder,
-                   DataframeRegisterCache& cache,
-                   IndicesReg indices,
-                   uint32_t row_count,
+  QueryPlanBuilder(BytecodeLowering& lowering,
                    const std::vector<std::shared_ptr<Column>>& columns,
                    const std::vector<Index>& indexes);
 
@@ -325,26 +274,23 @@ class QueryPlanBuilder {
   // accessed.
   void Output(const LimitSpec&, uint64_t cols_used_bitmap);
 
-  // Finalizes and returns the built query plan.
-  QueryPlanImpl Build() &&;
-
   // Processes non-string filter constraints.
   void NonStringConstraint(
       const FilterSpec& c,
-      const interpreter::NonStringType& type,
-      const interpreter::NonStringOp& op,
+      const NonStringType& type,
+      const NonStringOp& op,
       const interpreter::ReadHandle<interpreter::CastFilterValueResult>&
           result);
 
   // Processes string filter constraints.
   base::Status StringConstraint(
       const FilterSpec& c,
-      const interpreter::StringOp& op,
+      const StringOp& op,
       const interpreter::ReadHandle<interpreter::CastFilterValueResult>&
           result);
 
   // Processes null filter constraints.
-  void NullConstraint(const interpreter::NullOp&, FilterSpec&);
+  void NullConstraint(const NullOp&, FilterSpec&);
 
   // Processes constraints which can be handled with an index.
   void IndexConstraints(std::vector<FilterSpec>&,
@@ -356,125 +302,7 @@ class QueryPlanBuilder {
   // Returns true if the optimization was applied.
   bool TrySortedConstraint(FilterSpec& fs,
                            const StorageType& ct,
-                           const interpreter::NonNullOp& op);
-
-  // Whether the span being pruned holds the query's result rows, or only a
-  // temporary copy of them.
-  enum class NullPruneScope { kResultRows, kScratchOnly };
-
-  // Given a list of indices, prunes any indices that point to null rows
-  // in the given column. The indices are pruned in-place, and the
-  // `indices_register` is updated to contain only non-null indices.
-  void PruneNullIndices(uint32_t col,
-                        interpreter::RwHandle<Span<uint32_t>> indices,
-                        NullPruneScope scope);
-
-  // Given a list of table indices pointing to *only* non-null rows,
-  // if necessary, translates them to the storage indices for the given column.
-  // If no translation is needed, the indices are returned as-is.
-  // If translation *is* needed, the value of `in_place` determines
-  // whether the translation is done in-place or whether the data is stored
-  // in the scratch register.
-  //
-  // Returns a register handle to the translated indices (either
-  // `indices_register` or the scratch register).
-  interpreter::RwHandle<Span<uint32_t>> TranslateNonNullIndices(
-      uint32_t col,
-      interpreter::RwHandle<Span<uint32_t>> indices_register,
-      bool in_place);
-
-  // Ensures indices are stored in a Slab, converting from Range if necessary.
-  PERFETTO_NO_INLINE interpreter::RwHandle<Span<uint32_t>>
-  EnsureIndicesAreInSlab();
-
-  // Adds a new bytecode instruction of type T to the plan.
-  template <typename T>
-  T& AddOpcode(RowCountModifier rc);
-
-  // Adds a new bytecode instruction of type T with the given option value.
-  template <typename T>
-  T& AddOpcode(uint32_t option, RowCountModifier rc) {
-    return static_cast<T&>(AddRawOpcode(option, rc, T::kCost));
-  }
-
-  // Adds a new bytecode instruction of type T with the given option value.
-  template <typename T>
-  T& AddOpcode(uint32_t option, RowCountModifier rc, interpreter::Cost cost) {
-    return static_cast<T&>(AddRawOpcode(option, rc, cost));
-  }
-
-  PERFETTO_NO_INLINE interpreter::Bytecode&
-  AddRawOpcode(uint32_t option, RowCountModifier rc, interpreter::Cost cost);
-
-  // Sets the result to an empty set. Use when a filter guarantees no matches.
-  void SetGuaranteedToBeEmpty();
-
-  // Allocates a register for column data pointer and adds RegisterInit entry.
-  // Returns a HandleBase that can be assigned to typed data_register fields.
-  interpreter::RwHandle<interpreter::StoragePtr> StorageRegisterFor(
-      uint32_t col,
-      StorageType storage_type);
-
-  // Returns the index register for the given position.
-  interpreter::RwHandle<Span<uint32_t>> IndexRegisterFor(uint32_t pos);
-
-  // Returns the null bitvector register for the given column.
-  // For NonNull columns, returns an empty handle.
-  interpreter::ReadHandle<interpreter::NullBitvector> NullBitvectorRegisterFor(
-      uint32_t col);
-
-  // Returns the null bitvector register for the given column, ensuring a
-  // PrefixPopcount bytecode has been emitted for SparseNull columns.
-  // No-op for non-SparseNull columns or if already emitted.
-  interpreter::ReadHandle<interpreter::NullBitvector> EnsurePrefixPopcountFor(
-      uint32_t col);
-
-  // Returns the SmallValueEq bitvector register for the given column.
-  interpreter::ReadHandle<const BitVector*> SmallValueEqBvRegisterFor(
-      uint32_t col);
-
-  // Returns the SmallValueEq popcount register for the given column.
-  interpreter::ReadHandle<Span<const uint32_t>> SmallValueEqPopcountRegisterFor(
-      uint32_t col);
-
-  interpreter::ReadHandle<interpreter::CastFilterValueResult> CastFilterValue(
-      FilterSpec& c,
-      const StorageType& ct,
-      interpreter::NonNullOp non_null_op);
-
-  interpreter::RwHandle<Span<uint32_t>> GetOrCreateScratchSpanRegister(
-      uint32_t size);
-
-  // Alias for scratch register type.
-  using Scratch = interpreter::BytecodeBuilder::ScratchRegisters;
-
-  // Parameters for conversion to row layout.
-  struct RowLayoutParams {
-    // The column to be copied.
-    uint32_t column;
-
-    // Whether, instead of copying the string column, we should replace it
-    // with a rank of the string.
-    bool replace_string_with_rank = false;
-
-    // Whether the bits when copied should be inverted.
-    bool invert_copied_bits = false;
-  };
-  uint16_t CalculateRowLayoutStride(
-      const std::vector<RowLayoutParams>& row_layout_params);
-
-  interpreter::RwHandle<Slab<uint8_t>> CopyToRowLayout(
-      uint16_t row_stride,
-      interpreter::RwHandle<Span<uint32_t>> indices,
-      interpreter::ReadHandle<interpreter::StringIdToRankMap> rank_map,
-      const std::vector<RowLayoutParams>& row_layout_params);
-
-  void MaybeReleaseScratchSpanRegister();
-
-  void AddLinearFilterEqBytecode(
-      const FilterSpec&,
-      const interpreter::ReadHandle<interpreter::CastFilterValueResult>&,
-      const interpreter::NonIdStorageType&);
+                           const NonNullOp& op);
 
   bool CanUseMinMaxOptimization(const std::vector<SortSpec>&, const LimitSpec&);
 
@@ -486,29 +314,8 @@ class QueryPlanBuilder {
   // Reference to the indexes available.
   const std::vector<Index>& indexes_;
 
-  // The query plan being built.
-  QueryPlanImpl plan_;
-
-  // Current register holding the set of matching indices.
-  IndicesReg indices_reg_;
-
-  // Low-level bytecode builder for register allocation, bytecode storage,
-  // and scratch management.
-  interpreter::BytecodeBuilder& builder_;
-
-  // Register cache for caching column/index registers across Filter() calls.
-  DataframeRegisterCache& cache_;
-
-  // Tracks which columns have had PrefixPopcount bytecode emitted.
-  base::FlatHashMap<uint32_t, bool> prefix_popcount_emitted_;
-
-  // Last scratch registers returned by GetOrCreateScratchSpanRegister.
-  std::optional<Scratch> scratch_;
-
-  // Row count before the first selective (equality/IN) filter was applied. Used
-  // to avoid compounding the selectivity of multiple such filters: only the
-  // most selective one determines the estimate.
-  std::optional<uint32_t> selective_filter_base_row_count_;
+  // Turns the decisions made here into bytecode.
+  BytecodeLowering& lowering_;
 };
 
 }  // namespace perfetto::trace_processor::core::dataframe
